+++ src/Maintenance/Maintenance/ChangeLog.md
@@ -21,7 +21,7 @@
 ## Upcoming Release
 
 ## Version 1.7.0
-* Added support to acknowledge ScheduledEvents 
+* Added support to acknowledge ScheduledEvents using 'Set-AzScheduledEvents' and 'Set-AzScheduledEventsList' 
 
 ## Version 1.6.0
 * Added ChangeSafety Support